Initial Payments implementation
Add Payment::Account, Payment::Ledger,
and Payment::BookTransaction.
Together, these model the core types
that we need to move money around our system.
Getting money into and off of the platform
will be for the next PR.

When a trip is ended,
we create 'book transactions' that move money
from the customer's eligible ledgers.
'Eligible' here means those those that are associated
with vendor service categories that overlap
with the service's categories.
Also, categories are now hierarchical.

So for example, if a resident has a ledger with the
'grocery' category, and they purchase a service with
the 'organic grocery' category (a child of 'grocery'),
they can use this ledger to pay, in addition to
any other qualifying ledgers.

However, if they purchase a 'food' service,
the ledger is not eligible, because it can only be used
for 'grocery'.

Note that the eligibility hierarchy means that,
if they have a 'food' ledger, it can be used for
food/grocery/organic. If they have an 'organic' ledger,
it can only be used for 'organic', not food/grocery.

Book transactions move money between two ledgers;
in most cases, like in the case of paying for a trip,
this means moving money to or from the 'platform' payment
account ledgers.

When a customer is created, we create a 'cash' ledger automatically;
we also add this to the vendor services we create
during bootstrapping.

We dynamically create platform account ledgers as needed;
the category hierarchy means we can pull from platform ledgers
appropriately- for example, if we need to move money
from the platform 'food' ledger to a customer's 'food' ledger,
we can pull funds from the platform 'food' ledger and
all child ledgers that have a category assigned
that is a child of the 'food' category.

The next steps for this are:
- Expose resident ledgers through the dashboard.
- 'Funding' transactions using platform funds, like "Free $5 on signup"
- Manage funding methods (linked bank accounts),
  and use those to move money onto the platform.