Update current member in other API calls
Fixes https://github.com/lithictech/suma/issues/104

In many places, we update fields about the current member,
like when we add funds (removing read-only mode)
or start/end a trip (changing ongoing_trip).

We want to avoid re-fetching `/v1/me` when this happens,
since it's an extra API call, and we want to avoid
any additional API calls.

Endpoints can now call `add_current_member_header`,
which will set the `Suma-Current-Member` header,
which is a base64 encoded version of the 'current member' entity.

The on the frontend, we can use `tap(handleUpdateCurrentMember)`,
which will look for that header, decode and parse the value,
and update the stored user to it.

So now, when you begin/end a trip, the user gets updated along
with it.

This also adds caching of the current user into local storage,
optimistically assuming the user won't change when fetched
so we do not have to block rendering the page while fetching `/v1/me`.