Add language switcher, overhaul i18next init
Remove language detector stuff from i18next,
and store it ourselves in local storage.
We need to replicate some of what react-i18next is doing
with a context, so it's easier to inhouse this
(I am not sure there's much benefit using the detector
anyway).

Add a language switcher component.
Overhaul nav bar so it has a collapse icon.
Put the sudo and logout actions under it.