Add new 'waitlist' feature, and key-value table
Add endpoints for joining a waitlist.

This uses an overly-generic `member_key_values` table
which can store any key and value.
It should rarely ever be used,
but this sort of unstructured data is needed very early on,
so here it is.

Hook up UI to endpoints.

Add an Animated Checkmark component.