Add 'footer' into nav
Add what is traditionally a site footer, but put it into the
collapsable nav, which gets it out of the way for the normal app,
since it is not really a thing we need to show all the time.
It links to the privacy policy, terms, and has a contact email link.

Fixes https://github.com/lithictech/suma/issues/202