Fix bug after deleting cart items
If a cart item was deleted while it pointed to a checkout,
it would error due to a non-null constraint.

Instead we need to delete the checkout item entirely.
This however can lead to zero-item checkouts;
in these cases, we 403 when trying to fetch them,
since they are as good as deleted.