Automatically add subsidy when a member reigsters
Create an 'automation trigger' system
that does some of what the async job system does.

Add an implementation (and hook it up to bootstrap)
so we create a 'holiday 2022 ledger' when a user signs up,
and load it with $80 in credit.

Note this ledger can only be used for the holiday 2022 products
due to vendor service category constraints.

Fixes #241