Fix bug with charging multiple ledgers
When we calculate how we are paying for a product,
we figure out each ledger's contribution by looking at its balance.
However, once the ledger 'pays' for a product, it should deduct
from its balance (but we're not ready to add a book transaction).

We now have a 'calculation context' that allows us to build up
changes to a ledger balance as we process all the products.