Payment methods are enabled/disabled via config
Certain Suma deployments may only support certain payment instruments-
for example, it may be easy to get set up with cards but difficult to
start using bank accounts, or perhaps this is an entirely unbanked instance
that only uses script. When a payment method is not enabled:

- Funding transactions using those instruments is disabled.
- Adding those instruments is disabled.
- They do not show up in a member's 'usable instruments'.
- They do not show in the UI.

Removing support for an instrument on an established instance is not
considered a 'safe' operation since a disabled instrument
may be linked to active checkouts, etc. If this is needed in the future,
we need to add better support for it.

Fixes https://github.com/lithictech/suma/issues/235