Add build var and sentry support to frontend
Previously, we have no good way to pass in environment variables
between backend and frontend; we'd have to set REACT_APP variables
in the environment, and they were passed through to the apps,
but this is annoying because in many cases we should only
need to set them once.

This replaces the webapp build process to go through a Rake task;
this rake task and an accompaying bash script will proxy through
some variables (Stripe and Sentry keys, for example)
as `REACT_APP` vars.

Also hooks up Sentry for use in the frontend via the logging system.
Update the logging system to be more consistent
with structured logging, which often accepts `(message, context)`
(previously the logger worked like the console logger).