Fix busted images
We weren't handling data urls consistently,
and added a `data:` prefix to to a data url.
This was handled by Chrome but not on Safari.
Be more precise about how we handle institution
logo urls (they are base64 bytes from Plaid,
but already data urls for the ones we load).