Remove _md suffix from localization files
Keeping track of _md suffix on localization resulted in too much
confusion when strings changed to include or not include markdown.
So I wanted to get rid of it.

Doing this required either a lot of cleverness around
when to use and not use _md between base and other locales,
or just doing away with it entirely. I chose the latter.

However this meant we had no good way to know when
something should use markdown or not.
I tried to use react-markdown to render just strings,
but I found out how complex it is.
The only realistic option was to switch to markdown-to-jsx,
which we wanted to do anyway, and was very easy to use.

When checking that we are using the right helper (t or md),
we now render the actual markdown, and log an error if
there is no transformation (ie, t should be used)
or there is (ie, md should be used).

This more robust checking does not depend on the programmer
keeping track of when to use md or t; they just use what
they think is the correct one, and respond to any warnings.
This did expose a lot of misuse of these helpers,
and I cleaned up what I saw.

As part of this I also did some cleanup to the privacy policy,
so it didn't need to copy things from localization.
Instead it can create its own `Lookup` instance
for a given strings namespace,
with its own `t` and `md` methods, and use that.

Fixes https://github.com/lithictech/suma/issues/315 (ignore _md),
and https://github.com/lithictech/suma/issues/314 (render entities)
and https://github.com/lithictech/suma/issues/307
(replace react-markdown).