Contact list/quick-enroll support (#391)
Add a page (`/contact-list`) where users can sign up with a name and email. For use during community signup events. 

Page can be given an `eventName` query param that is persisted across signups using the page, and is stored on the referral.

Adds a 'referral' database model.

Also replace `react-phone-number-input` with a custom mask.

---------

Co-authored-by: Rob Galanakis <rob.galanakis@gmail.com>