Add support for deep linking to mobility adapter
The 'mobility vendor adapter' system had support for
beginning and ending trips, but not deep linking.

This adds support for deep linking,
by returning a rental URL from the `GET /vehicle` endpoint
which is called before reserving a vehicle.

There's also support now for indicating if deep linking
can be used, or if the user must set up a private account first.

The reservation card is updated to support deep linking
and private account forwarding.

Adds a migration to `Vehicle`, and stores the `rental_uris` from the
GBFS feed, so they can be returned to the frontend in the case
of deep linking. Note that the appropriate deep link
is parsed based on the user agent platform.

Fixes #435