Add 'latest_access_code' to vendor accounts
Instead of showing recent text bodies,
we store the latest access code on the vendor account
and return it to the frontend if it's recent enough.

We need this because having a structured access code
is important since we can't send SMS.

Also adds a new `Copyable` component,
which can be used through Markdown.