Add endpoint for long polling of private accounts
Adds an endpoint,
`/v1/anon_proxy/vendor_accounts/poll_for_new_access_codes`,
which is used to long-poll for changes to
any vendor account 'last access code' for a member.

This long polling runs on the private accounts page only.

It's important we long rather than short poll because
we want to be as light as possible on the user's device.

In the future we should switch to web sockets,
but for now, long polling is sufficient.

Also adds a 'blinking-alert' class,
which we use to highlight the 'access code' alert
when it shows up.