Do not show ledger credits when they're ambiguous (#479)
If a product has less > 1 max quantity, do not show nonledger contributions on list and details page. It makes it confusing if we show it because it's not clear what the final price is (like $5 credit and $2 unit cost would show $0 cost). Since it's unambiguous with 1 max quantity, we leave it, but hide it when we can purchase multiple.

---------

Co-authored-by: Rob Galanakis <rob.galanakis@gmail.com>