Payout (refund) support (#495)
Implement the concept of 'payouts', detailed in the original payment system. Payouts represent money moving off the platform.

There are many reasons money can leave the platform, so payouts have different classifications, based on what is set in terms of their `refunding_funding_transaction`, `crediting_book_transaction`, and `originated_book_transaction`.

Originated book transaction represents the book transaction from the user's ledger to the platform ledger,
to represent that money is leaving their ledger and eventually the system.

Crediting book transaction represents the book transaction from the platform ledger to the user's ledger,
to compensate them for a refund.

The refunding funding transaction points to the funding transaction used in a refund, if any.

- 'refund' has a credit and refunding transaction,
  and is generally a reversal of money that was used to pay for something and charged during a purchase.
- 'reversal' has no credit but does have a refunding transaction,
  and is generally a reversal of money added to the ledger, not used directly in a purchase.
- 'payout' is money we send to a payment account, like to a vendor's bank account.
- 'platformpayout' is money we send off-platform without any associated book transaction.
- 'unknown' is a fallback and should not be seen with a valid combination of fields.

Refunds are currently created from refunds in Stripe using a scheduled job (which can be triggered from webhookdb). So admins can refund in Stripe and we'll pick it up automatically after a little while.

See `payments.md` for use case examples.