Create Mdx comp to use new translation hook
The new component must be a React component because it uses a useI18Next hook. useI18Next will likely be renamed.
Use this component to test on a localized key - it works, needs to take into account other methods we use e.g.
translating $t{strings:ns} inside of a localized string.
- Update compileStringAsync due to deprecated React.render function
- Clean up new 't' hook and complete some TODOs
- Add TODOs