Overhaul private account provisioning (#516)
Instead of a multi-step process with user intervention, treat private account provisioning as a 'headless' process that must involve following a magic link. The backend makes an API request to Lime to create a user, and then listens for a magic link sent to the user's anonymous proxy email. When it's found, we open the magic link, logging them into Lime.

There are also a couple other changes here:

- Translation fixes
- Remove hard requirement on geolocation services. If we can't locate in the browser, we fall back to IP geolocation. Since we aren't taking scooter rides from Suma, this is sufficient; once we are taking rides from suma (using an MaaS under the hood), we'll need actual location services; see #511 for the basis of that work.
---------

Co-authored-by: Denilson Velasquez <66847768+DeeTheDev@users.noreply.github.com>