Add a topic shim job
In some cases we want to use simpler events,
like `suma.member.verified`. This shim job looks at all jobs
and emits an event based on custom logic.

This is sort of a gross workaround,
but it's needed for automation triggers to avoid making them
way more complex. Once we get rid of automation triggers,
we should look at eliminating this job as well.