Send a private account login SMS when we receive a Lime auth email (#593)
Do not show the link in the UI since going to the url from the browser does not reliably launch the Lime app (it may work for other apps- we have confirmed this is a bug on the Lime side, with how it redirects/deep links). So send an SMS instead, which uses a different URL handling which seems to work more reliably.