Move calculation contexts to the API level
They should be passed in from the top,
so we can do things like the payment automation simulation
to generate adjustments in advance.

This also required making them immutable,
since if we use them multiple times we need them
to be idempotent.

Commerce cart and checkout now have a 'cost info' object
which collects a bunch of methods together using a single context.
This will allow much more efficient caching and cohesion
in the future when we need to run the subsidy simulations
using different contexts, and choose one to render entities with.