Require products and ledgers to have categories to calculate charges
Since products without categories cannot be used to find any sort of
contributions, it'd always be cash; this is almost never what we want,
so always require a VSC for products.

For ledgers, having no categories means it cannot be used for ANYTHING,
which would never be what we want, so require a VSC for ledgers.