Remove PG ModelUtilities assumptions
We need to add a new Model type,
and want to use some of the same functionality
in ModelUtilities, but not all of it.
For example, we don't want all models to have pubsub behavior.

This change pulls out a new `ModelPubsub` module
to mix in for pubsub behavior, and also moves some things
like automatically loaded plugins from `ModelUtilities`
over to the model itself, unless the utilities actually
depend on the plugin (like `plugin(:subclasses)`).

Also removes some unused methods.