Member enrollment organization affiliation (#624)
FIxes #546 

It would be helpful for suma staff to have a way to know what affiliated organizations members are eligible for when they signup. This change models Organizations and OrganizationMemberships, the organizations represent entities like "Hacienda CDC" or affordable housing partners, the membership represents the member who is affiliated with that organization.

When members signup regularly or through the contact list add page, we show them available organizations, they can choose one, we create a membership based on the organization chosen. If an organization does not exist, we allow them to type their own and we add a summary with that org so that suma staff can verify it in the adminapp. This helps expedite the member signup process for suma staff.

- Add model migrations and ensure that Organization Memberships are unique and not duplicated
- Add Organization & Organization Membership fixtures
- Setup adminapp CRUD routes for `Organization` and `Organization Memberships`
- Add ability to update/create/remove member organization memberships in the `MemberEditPage`
- Render memberships in the `MemberDetailPage`