Update lime email content (#665)
We are fetching new emails from lime webhooks but we aren't matching the message content correctly for these new emails contents. Improve the matching for the token and link in the email contents. Add tests to support this.

---

Also fix exposed variables in `VendorAccountMessageEntity` and fix display in the frontend.

---------

Co-authored-by: Rob Galanakis <rob.galanakis@gmail.com>