Add a url shortener app, hook it up to Lime
SMS on Android are suddenly replacing links like `?magic_link=123`
with `?magic link=123` (underscores to spaces), which is breaking
link parsing in messaging/SMS apps causing the link to be unclickable.

This change solves that by using a url shortener,
so that `https://app.mysuna.org/r/abc123`
goes to the original Lime url.

This adds a new `UrlShortener` application
(which is not tied to Suma directly) to encapsulate
a simple url shortner which can support a pretty decent throughput.
Note that because we want to eliminate 3rd party dependencies
where feasible, especially around privacy, hosting costs,
and provisioning simplicity, this was simpler to build than integrate.
The behavior, limitations, and implementation are
documented in the code and aren't worth repeating here.