Configure image blobs in CSP builder (#681)
Trying to upload and display an image in admin caused broken image since the content security policy was not allowing image  blob:. This change adds the img_blob setting to the CSP Policy builder allow blobs to be loaded and displayed. Then allow this setting in the admin app (not necessary for the normal app).