Rework ledger lines with recent lines
We want to show recent lines first when there are ledger transactions.
And we want to show the first ledger when there are no transactions.

Backend:
- Rework ledgers view to expose more line details to the API
- Remove unnecessary items in `Suma::Member::Dashboard` and add necessary to `Suma::Payment::LedgersView` like `lifetime_savings`
- Update/add tests

Frontend:
Rework ledgers since there are no longer first page transactions.
Display recent lines when there are transactions available.
Otherwise we render the first ledger (which should not have lines,
so we avoid further API calls).
If a ledgerParamId is available, we render that ledger information
and fetch their lines.
If there is a ledgerParamId, theres no active ledger and theres
ledger lines, render error page since the ledger couldn't doesn't exist.