Add resource-based role access
In the frontend, we need to match the role access for a feature.
So the API and frontend need to use the same knowledge of
'members use the :admin_members role access'.

Build a mapping of resource to role access,
like `Member => [:admin_member, :admin_management]`,
and expose it to the frontend.
Also use this to control the common endpoints,
rather than the endpoint declaring what access it uses
(it does still need to match in one-off endpoints).

Add `useRoleAccess` and `useResourceAccess` hooks,
and hook them up to list and detail pages
(and a few others) to control what is visible.