Refactor i18n rewrite to support nesting better
We need to handle nested localizations, where:
- `"abc": "**hi**"`, uses the `m` formatter, and
- `"s1": "$t(abc)"`, depends on `abc`,
  but doesn't know it will need markdown.

So we do a first pass to find all the 'naive' formatters,
and then another pass where we look at all the formatters
for all the (recursive) nestings, so we'd know that `s1` gets `m`
in our example.

This required a pretty big refactor, with internal classes
to keep track of the nestings, and now we have more friendly
formatter classes. But fundamentally everything works
the same as the previous commit (and the output format
didn't change at all).