Analytics models: charge, book/payout/funding transactions
Fixes the ledger model denormalization to ALL ledgers
show up in the table. What was there, was dumb,
stomping on the same ledger for every book transaction.

Add models for charges and book, payout, and funding transactions.

We may wish to denormalize things further in the future.
Useful queries are probably still going to require
some JOINs on other analytics tables.