Program concept overhaul (#726)
Implement a Program and Enrollment concept based on the PWB pilot. The backend logic of member resource eligibility is replaced with, 'a user can access an offering/service if the user has an approved program enrollment (or the user is a verified member of an organization with an approved program enrollment) in an active program that is associated with the offering/service'.

- `Vendible::Grouping` dashboard list is replaced by `Program` detail list that a member is enrolled in.
- offerings/vendor configs/vendor services/triggers `Eligibility::Constraint` associations are replaced with `Programs`
- members `Eligibility::Constraint` associations are replaced with `Program::Enrollment`, and organizations can be associated with program enrollments too
- remove all backend eligibility constraint models and references

Sets a 'now' key and API method helper in the env which we can use across the API and model layer of requests. This avoids many different definitions of 'now' within an endpoint/entity.

Admin pages replace eligibility constraints with programs and program enrollments. Also removes 'vendible groups', not necessary anymore.

Members with a positive ledger cash balance, will see their available amount on the top of the dashboard page

---------

Co-authored-by: Rob Galanakis <rob.galanakis@gmail.com>