Dynamic translations can use markdown
I18n formatters can hide unicode zero-width flags at the start
of translated strings, which will determine what formatter
to use on the frontend.

Allows us to continue to use `{product.description}`
but replace it with `{dt(product.description)}` if we want to
support markdown in a component.

The frontend cost of this is very minor (checking `str[0]`).
The backend cost is nontrivial (we have to render markdown
for every translated text), but it's fast enough,
and we have an LRU so we don't need to keep re-testing
strings that we repeatedly re-render
(in most cases, just a few hundred strings will be shown in
the app at a given time).

Support markdown in the program and product descriptions,
use multiline for some 'localized text' inputs in admin,
and fix a missing exposure of vendor configuration instructions.