Resources render the last image vs the first
Since we are not displaying multiple images in the
front facing app yet, we can expose the last image
as the featured image. Same with admin, when they
upload an image for a resource, that image will be
displayed in the front facing app.

Note that we may add an image slider sooner or later.