Add charge line items
docs from charges class:

Charges represent something a member was charged for,
on or off platform. They are *not* part of the payment system;
they are a higher-level representation of a charge,
linked to something like a commerce order or mobility trip.

Each charge may have one or more 'line items'.
Each line item represents some part of the charge.

Line items may represent on-platform funds flows (book transactions),
like the money moved from each ledger to pay for an order
that was paid from multiple ledgers (cash, subsidy, etc.).

Line items may also represent off-platform funds flows,
where there are no funds flows to represent. These are 'self'
line items, which do not point to a book transaction
(each line item can have 'self' data, or point to a book transaction).
An example would be when we have a user account linked to
an external vendor like Lyft, and Lyft handles all charging;
suma is there as an intermediary but is not in the funds flow.
Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.