Fix bug with i18n format
Strings refererences were like `$t(strings:x)`,
but the nested strings just had `x`.
We need to always have a namespace when processing strings.
This caused a bug where the `strings:faq` string would use
the `md` formatter, but a string that refrred to it
via `$t(strings:faq)`, would not find the nesting,
so would not pick up the markdown.

Also modify the mobility map intro text,
since Ride Connection and Lime are programs now.

There was also a latent bug with cross-resource references;
we need to 'prime' all resource files, not just a single file.