Member activities are more of an audit system
Add a helper so subjects of audit activities
can also list their audit activities.

Add a generated column and index of
the activity subject id as an integer.

We need this for org roles, to keep track of who changed the roles,
just like we do for members.

fixup has audit activity