Replace map cards with drawer
The cards were underlying over the new drawer.
We couldn't just change their location, because the drawer
could change size, and the card could get pushed so low
that it could be near the center of the screen
(where the scooter is).

Instead, all content is now in the drawer, and the drawer
resizes as neccessary.

Also add a new 'internal' mobility adapter
we can use for MaaS-like testing, so we can explore
the additional card states.