Fix bug fetching lyft rides (or authing)
The lyft program ID needs to come from the URL.
However that ID is NOT what's used to fetch rides.

Instead, use an endpoint to get the 'account id' (used to fetch rides)
from the lyft program id.

NOTE: The wrong program ID will error when enrolling users,
but will miss any rides when fetching rides.
This is why we didn't notice that there were two, not just one,
ids in use.