Use a model for GBFS feeds
We cannot just configure gbfs feeds, since we may need
many feeds for the same vendor, and there was a lot
of implicit dependencies on specific vendors and names.

Instead, add a GbfsFeed model that declares the root url,
and which of the components (free bike status, geofencing zones)
to sync, and when they were synced.

Rework all the code to remove the implicit dependencies on lyft
and lime vendors, except for the deeplinking where depending on
a vendor slug (but not creating the vendor) makes sense.