Automatically reload when asset js not found
Instead of a broken app with a missing asset,
reload the page.

When loading the app, make a HEAD to the hashed JS file.
If it isn't found, it's probably because the site has been
redeployed. If this happens, automatically reload the entire page,
using 'clearcache=(now)' to bust the recent cache
(use Date.now(), not just 'true', to make sure this works
even in the face of rapid deploys).

If the index still fails to load, maybe because the entire site
is down, show an error message.

This requires:

- Writing out the build manifest
- Parsing the manifest at runtime, to load the hashed index.js path.
  vite does not make it available to template into index.html.
- Writing this hashed file path into dynamic config.
- An inline JS function as described above.

Also fix an issue where a 404 on a JS or CSS file,
would return as a valid HTML file which would of course
fail to parse.