Improve app nav with scaled fonts
Enable pinch zooming on the web app.

Fix the top nav so that elements do not try to lay out normally;
this causes them to break and stack, which isn't suitable for the nav.
Instead, use absolute positioning so they are aligned to the right,
even if they have to stack on each other.
The need for this is clear when the offline nav indicator is visible.