Add external links and admin actions to messaging (#819)
The admin-facing changes here add external links to Message Deliveries, adds a new 'admin actions' concept, which for messages can fetch the data from the API directly so we can see what's up (can improve this in the future), and adds a link from reset codes to their delivery.

However, doing this ran into massive problems with the way messaging transports were designed. Each transport (sms, email) also was tied to a carrier (signalwire, postmark). Except sms was sometimes using twilio verify. We added some gymnastics to make this work, but then it required another level of gymnastics to unwind it, so a delivery could know "use the link to twilio verify instead of signalwire for this message".

The real solution here was to separate the concept of 'carrier' (signalwire, twilio sms, twilio verify, postmark) and 'transport' (sms, email, etc- the semantics common to all carriers for a transport). This means the transport can have a 'current' carrier, but old carriers could still be supported. For example, if we switched from signalwire sms to someone else, the signalwire carrier could stick around, but the sms transport would start using a new carrier.

This also meant adding a new 'sms otp' transport, since the semantics of it are different from normal sms (we don't send a body, just a code). This can use a generic 'twilio verify' carrier. If we added WhatsApp verification for example, we'd need a new transport for it (`whatsapp_otp`), but not a new carrier.

This messaging code is among the oldest in the app, not remodeled substantially from the Cozy days. But suma has introduced a new set of needs since it uses primarily SMS, which are more complex than emails.

- Add `Message::Carrier` classes
- Move and rename `Message::Transport` classes
- Add `Suma::AdminActions`
- Rename `delivery.transport_service` to `delivery.carrier_key`.